# Conflicts:
#	src/main/java/com/lit_map_BackEnd/domain/work/service/VersionService.java
#	src/main/java/com/lit_map_BackEnd/domain/work/service/VersionServiceImpl.java

+++ src/main/java/com/lit_map_BackEnd/domain/work/service/VersionServiceImpl.java
@@ -10,6 +10,7 @@ import com.lit_map_BackEnd.domain.character.repository.CastRepository;
 import com.lit_map_BackEnd.domain.character.service.CastService;
 import com.lit_map_BackEnd.domain.mail.dto.MailWorkDto;
 import com.lit_map_BackEnd.domain.mail.service.MailService;
+import com.lit_map_BackEnd.domain.member.entity.MemberRoleStatus;
 import com.lit_map_BackEnd.domain.work.dto.VersionListDto;
 import com.lit_map_BackEnd.domain.work.dto.VersionResponseDto;
 import com.lit_map_BackEnd.domain.work.entity.*;
@@ -24,6 +25,7 @@ import com.lit_map_BackEnd.domain.member.repository.MemberRepository;
 import com.lit_map_BackEnd.domain.work.entity.Confirm;
 import com.lit_map_BackEnd.domain.work.entity.Version;
 import com.lit_map_BackEnd.domain.work.entity.Work;
+import org.springframework.security.core.Authentication;
 
 
 import java.util.*;
@@ -89,11 +91,18 @@ public class VersionServiceImpl implements VersionService{
 
     @Override
     @Transactional
-    public void deleteVersion(Long workId, Double versionNum) {
+    public void deleteVersion(Member member, Long workId, Double versionNum) {
         Work work = workRepository.findById(workId)
                 .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.WORK_NOT_FOUND));
 
-        versionRepository.deleteByWorkAndVersionNum(work, versionNum);
+        MemberRoleStatus writer = work.getMember().getMemberRoleStatus();
+        MemberRoleStatus memberRoleStatus = member.getMemberRoleStatus();
+        if (memberRoleStatus.equals(writer) || member.getMemberRoleStatus() == MemberRoleStatus.ADMIN) {
+            versionRepository.deleteByWorkAndVersionNum(work, versionNum);
+        } else {
+            throw new BusinessExceptionHandler(ErrorCode.WRITER_WRONG);
+        }
+
     }
 
     @Override
@@ -133,7 +142,9 @@ public class VersionServiceImpl implements VersionService{
 
     @Override
     @Transactional
-    public void rollBackDataSave(Long workId, Double versionNum) {
+    public void rollBackDataSave(Long memberId, Long workId, Double versionNum) {
+        memberRepository.findById(memberId)
+                .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.WRITER_WRONG));
         Work work = workRepository.findById(workId)
                 .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.WORK_NOT_FOUND));
         Version version = versionRepository.findByVersionNumAndWork(versionNum, work);
@@ -178,27 +189,6 @@ public class VersionServiceImpl implements VersionService{
     public MailWorkDto sendMailWithTemplate(Long versionId, String subject, String content) {
         Version version = versionRepository.findById(versionId)
                 .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.VERSION_NOT_FOUND));
-
-        Work work = workRepository.findById(version.getWork().getId())
-                .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.WORK_NOT_FOUND));
-
-        Member member = memberRepository.findById(work.getMember().getId())
-                .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.USER_NOT_FOUND));
-
-        String litmapEmail = member.getLitmapEmail();
-
-        mailService.sendEmail(litmapEmail, subject, content);
-
-        return new MailWorkDto(litmapEmail, work);
-    }
-
-    public void approveMail(Long versionId) {
-
-        if (!adminAuthService.isAdmin()) {
-            throw new BusinessExceptionHandler(ErrorCode.FORBIDDEN_ERROR);
-        }
-            Version version = versionRepository.findById(versionId)
-                    .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.VERSION_NOT_FOUND));
 
             Work work = workRepository.findById(version.getWork().getId())
                     .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.WORK_NOT_FOUND));
